A first stab in low power mode for the MPU6050
I'm not super happy with it. The MPU is put to sleep and the DMP is
disabled. When they are turned on, it seems to take ~5 s for the
orientation to converge.

Another possibility is to leave it running and lower the sample rate,
but I have to do some measurements to understand how power hungry  and
efficient that would be.